feat: prepare for uniform editor
