+++ assets/test_project.bulin
@@ -1,1 +1,1 @@
-{"file":"/home/nickel/Programming/repos/bulin_rust/assets/test_shader.wgsl","content":"@fragment\nfn fs_main(@builtin(position) position: vec4<f32>) -> @location(0) vec4<f32> {\n    let time = 0.;\n    let radius = 0.5;\n    let leaves = 7.0;\n    let speed = 0.5;\n    let flower_color = vec3<f32>(0.5, 0.1, 0.2);\n    let test = vec3<f32>(0.1, 0.2, 0.5);\n\n    let st = (position.xy - uniforms.position) / uniforms.resolution;\n\n    let d = 0.25 + 0.5 * (0.5 + 0.5 * sin(2.0 * time));\n    let pos = vec2<f32>(d) - st;\n\n    let r = (2.0 + sin(9.0 * time)) / radius * length(pos) * 1.2;\n    let a = atan2(pos.y, pos.x);\n\n    let f = abs(cos(a * leaves / 2.0 + speed * time)) * 0.5 + 0.3;\n    let ratio = smoothstep(f, f + 0.02, r);\n    let color = flower_color * (1.0 - ratio);\n    let back = test * ratio;\n\n    return vec4<f32>(color + back, 1.0);\n}\n","theme":"SolarizedDark"}
+{"text_editor":{"file":"/home/nickel/Programming/repos/bulin_rust/assets/test_shader.wgsl","content":"@fragment\nfn fs_main(@builtin(position) position: vec4<f32>) -> @location(0) vec4<f32> {\n    let time = 0.;\n    let radius = 0.3;\n    let leaves = 7.0;\n    let speed = 0.5;\n    let flower_color = vec3<f32>(0.5, 0.1, 0.2);\n    let test = vec3<f32>(0.1, 0.2, 0.5);\n\n    let st = (position.xy - uniforms.position) / uniforms.resolution;\n\n    let d = 0.25 + 0.5 * (0.5 + 0.5 * sin(2.0 * time));\n    let pos = vec2<f32>(d) - st;\n\n    let r = (2.0 + sin(9.0 * time)) / radius * length(pos) * 1.2;\n    let a = atan2(pos.y, pos.x);\n\n    let f = abs(cos(a * leaves / 2.0 + speed * time)) * 0.5 + 0.3;\n    let ratio = smoothstep(f, f + 0.02, r);\n    let color = flower_color * (1.0 - ratio);\n    let back = test * ratio;\n\n    return vec4<f32>(color + back, 1.0);\n}\n","theme":"SolarizedDark"},"uniforms_editor":{"uniforms":[],"current_add_candidate":{"value":"Int","name":""},"time":null}}
