+++ src/text_editor/highlighter.rs
